Finally figured out how to pass a mutable reference lol

+++ rust/game-of-life/src/cell.rs
@@ -1,10 +1,11 @@
-const NEIGHBOURS_TO_LIVE: u32  = 2;
-const NEIGHBOURS_TO_PROCREATE: u32  = 3;
-const NEIGHBOURS_TO_DIE: u32  = 4;
+const NEIGHBOURS_TO_LIVE: u32 = 2;
+const NEIGHBOURS_TO_PROCREATE: u32 = 3;
+const NEIGHBOURS_TO_DIE: u32 = 4;
 
-#[derive(Debug)]
-pub enum Cell{
-    Alive { x: i32, y: i32, neighbours: i32 },
-    Baby { x: i32, y: i32, neighbours: i32 },
-    Empty
+#[derive(Debug, Copy, Clone)]
+pub enum Cell {
+    Alive { x: usize, y: usize, neighbours: i32 },
+    Baby { x: usize, y: usize, neighbours: i32 },
+    Empty,
 }
+
